% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{BostonHousing}
\alias{BostonHousing}
\title{BostonHousing data set.}
\format{A data frame with lots of rows and variables:}
\usage{
BostonHousing
}
\description{
Describe the data set
}
\keyword{datasets}

