% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/caridge.R
\name{caridge}
\alias{caridge}
\title{Caret Ridge BostonHousing}
\usage{
caridge(X)
}
\arguments{
\item{BostonHousing}{only works with this data set at the moment.}
}
\value{
\code{caridge} returns an object of class "\code{ridge}".
}
\description{
Use \code{caridge} to perform ridge regression.
}
\examples{
caridge(BostonHousing)

\dontrun{
caridge(TRUE, TRUE)
}
}
\seealso{
\code{\link{lm}}, \code{\link{class}}, \code{\link{formula}}
}

