% Generated by roxygen2: do not edit by hand
<<<<<<< HEAD
% Please edit documentation in R/linreg.R
\name{linreg}
\alias{linreg}
\title{Perform ridge regression}
=======
% Please edit documentation in R/linreg-wrapper.R, R/linreg.R
\docType{package}
\name{linreg}
\alias{linreg}
\alias{linreg-package}
\title{Perform linear regression}
>>>>>>> ca9124b600e1deff0be5eb5654b44da44c7e7807
\usage{
linreg(formula, data)
}
\arguments{
\item{formula}{an object of \code{\link{class}} \code{\link{formula}} (or one
that can be coerced to that class): a symbolic description of the model to
be fitted.}

\item{data}{a data frame or (or object coercible by
\code{\link{as.data.frame}}).}
<<<<<<< HEAD

\item{lambda}{a scalar or vector of ridge constants}
}
\value{
\code{ridgereg} returns an object of class "\code{Linreg}".

  An object of class "\code{Ridgereg}" have the following methods:
  \item{coef()}{a named vector of coefficients}
  \item{pred()}{predicted values of of a model}
}
\description{
Use \code{ridgereg} to perform ridge regressions.
=======
}
\value{
\code{linreg} returns an object of class "\code{Linreg}".

  An object of class "\code{Linreg}" have the following methods:
  \item{coef()}{a named vector of coefficients}
  \item{resid()}{the residuals, that is response minus fitted values}
  \item{pred()}{predicted values of of a model}
  \item{print()}{print the model}
  \item{plot()}{plot the residuals of the model}
  \item{summary()}{summary of the model esitmation}
}
\description{
Use \code{linreg} to perform linear regressions.

The package provides six functions to manipulate linear regression
models within a new class {\code{LinregClass}}: \itemize{
\item{\code{\link{coef}} returns the coefficients as a named vector.}
\item{\code{\link{resid}} returns the vector of residuals e.}
\item{\code{pred} returns the predicted values.}
\item{\code{\link{print}} prints the coefficients and coefficient names.}
\item{\code{\link{plot}} creates two plots, residuals vs. fitted values and
square root of residuals vs. fitted values.}
\item{\code{\link{summary}} returns the coefficients with their standard
error, t-value, p-value, estimated standard deviation and degrees of freedom.} }
>>>>>>> ca9124b600e1deff0be5eb5654b44da44c7e7807
}
\examples{
linreg(formula = Sepal.Length ~ Sepal.Width, data = iris)

\dontrun{
linreg(TRUE, TRUE)
}
}
\seealso{
<<<<<<< HEAD
\code{\link{lm.ridge}}, \code{\link{class}}, \code{\link{formula}}
=======
\code{\link{lm}}, \code{\link{class}}, \code{\link{formula}}
>>>>>>> ca9124b600e1deff0be5eb5654b44da44c7e7807
}

