% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LiuLinReg.R
\docType{package}
\name{LiuLinReg}
\alias{LiuLinReg}
\alias{LiuLinReg-package}
\title{LiuRLab3: A package for laboratory work 3 of the Advanced R Programming
course.}
\description{
The package provides six functions to manipulate linear regression
models within a new class {\code{LinregClass}}: \itemize{
\item{\code{\link{coef}} returns the coefficients as a named vector.}
\item{\code{\link{resid}} returns the vector of residuals e.}
\item{\code{pred} returns the predicted values.}
\item{\code{\link{print}} prints the coefficients and coefficient names.}
\item{\code{\link{plot}} creates two plots, residuals vs. fitted values and
square root of residuals vs. fitted values.}
\item{\code{\link{summary}} returns the coefficients with their standard
error, t-value, p-value, estimated standard deviation and degrees of freedom.} }
}

