% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ridgereg.R
\name{ridgereg}
\alias{ridgereg}
\title{Perform ridge regression}
\usage{
ridgereg(formula, data, lambda)
}
\arguments{
\item{formula}{an object of \code{\link{class}} \code{\link{formula}} (or one
that can be coerced to that class): a symbolic description of the model to
be fitted.}

\item{data}{a data frame or (or object coercible by
\code{\link{as.data.frame}}).}

\item{lambda}{lambda of a model}
}
\value{
\code{ridgereg} returns an object of class "\code{Ridgereg}".

  An object of class "\code{Ridgereg}" have the following methods:
  \item{coef()}{a named vector of coefficients}
  \item{resid()}{the residuals, that is response minus fitted values}
  \item{pred()}{predicted values of of a model}
}
\description{
Use \code{ridgereg} to perform ridge regressions.
}
\examples{
linreg(formula = Sepal.Length ~ Sepal.Width, data = iris)

\dontrun{
linreg(TRUE, TRUE)
}
}
\seealso{
\code{\link{lm}}, \code{\link{class}}, \code{\link{formula}}
}

