% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/airport_delay.R
\name{visualize_airport_delay}
\alias{visualize_airport_delay}
\title{Visualize Airport Delays}
\usage{
visualize_airport_delay()
}
\value{
ggplot object displaying the mean delay of flights for different airports
}
\description{
This function uses the data set \code{nycflights13} of the
\code{nycflights13} package to visualize the mean delay of
flights for different airports by longitude and latitude.
}

