% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize_airport_delays.R
\name{visualize_airport_delays}
\alias{visualize_airport_delays}
\title{Average departure and arrival delays}
\usage{
visualize_airport_delays()
}
\value{
A ggplot2 object
}
\description{
The function calculates the average departure and arrival
delays of of all airports of the dataset \code{nycflights13}.
The function then prints a map of the United States of America and
visualizes the average delay per airport in a point diagram.
The size of the points represents the average delay time for the
specific airport.
}

